requirements.txt
datadog-api-client>=2.28.0
anthropic>=0.40.0
python-dotenv>=1.0.0
certifi>=2024.2.2
google-adk>=0.4.0
litellm>=1.50.0
deepchecks-llm-client[otel]>=0.36.0
